Add a "-l/--load-sysinfo" option for "config reload" to merge the system info on device
The information includes:
-- platform string
-- system mac
-- port configuration: alias, lanes, speed, index

This is useful when we first load the configuration from config_db.json generated
outside which does not have the information on the device.

The default behaviour is not changed, so it won't impact the system unless you use this option.

This can be used for ZTP to config the box the first time and avoid any addtional work
which could cause inconsistency.